requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
requests>=2.31.0
plotly==5.23.0
fastapi==0.112.0
uvicorn==0.30.5
python-dotenv==1.0.1
